requirements.txt
# Bitget REST + Telegram sendMessage — обидва через requests
python-dotenv>=1.0.0
requests>=2.31.0
